[Refactor: To get user email from token as a UserID in notes]

+++ FundoNotes/src/services/notes.service.js
@@ -8,22 +8,23 @@ export const createNote = async (body) => {
 };
 
 //get all notes
-export const getAllNotes = async () => {
-  const data = await Notes.find();
+export const getAllNotes = async (UserID) => {
+  const data = await Notes.find({ UserID });
   return data;
 };
 
 //get a note by id
-export const GetNote = async (id) => {
-  const data = await Notes.findById(id);
+export const GetNote = async (id, UserID) => {
+  const data = await Notes.findById({ _id: id, UserID: UserID });
   return data;
 };
 
 //update a note
-export const updateNote = async (_id, body) => {
+export const updateNote = async (id, UserID, body) => {
   const data = await Notes.findByIdAndUpdate(
     {
-      _id
+      _id: id,
+      UserID: UserID
     },
     body,
     {
@@ -34,14 +35,14 @@ export const updateNote = async (_id, body) => {
 };
 
 //delete a Single note
-export const deleteNote = async (id) => {
-  await Notes.findByIdAndDelete(id);
+export const deleteNote = async (id, UserID) => {
+  await Notes.findByIdAndDelete({ _id: id, UserID: UserID });
   return '';
 };
 
 //archieve a note
-export const archiveNote = async (_id) => {
-  const note = await Notes.findOne({ _id: _id });
+export const archiveNote = async (id, UserID) => {
+  const note = await Notes.findOne({ _id: id, UserId: UserID });
   const isArchived = note.isArchived === false ? true : false;
   const data = await Notes.findByIdAndUpdate(
     {
@@ -56,8 +57,9 @@ export const archiveNote = async (_id) => {
 };
 
 //trash a note
-export const trashNote = async (_id) => {
-  const note = await Notes.findOne({ _id: _id });
+export const trashNote = async (id, UserID) => {
+  const note = await Notes.findOne({ _id: id,
+    UserID: UserID });
   const isTrash = note.isTrash === false ? true : false;
   const data = await Notes.findByIdAndUpdate(
     {
